Soft assertions to the rescue org.testng.SoftAssert
Create new SoftAssert() and then use it as in AssertJ:

    SoftAssert soft = new SoftAssert();
    //... test code here, then:
    soft.assertTrue(...);
    soft.assertEquals(...);
    soft.assertAll();

Now IDE reports all things, like Maven does.

Added Miriamele, so Seoman won't be lonely. :-) And so we have someone
of age in Europe but not in US, as I gave strict inequality in
conditions and forgot about it.

The use of COBOL cripples the mind; its teaching should therefore be regarded as a criminal offense.
      -- E.W. Dijkstra